# admin-go-api/docs/swagger.yaml
definitions:
  entity.LoginDto:
    properties:
      idKey:
        description: UUID
        type: string
      image:
        description: 验证码
        maxLength: 6
        minLength: 4
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - idKey
    - image
    - password
    - username
    type: object
  entity.SysPost:
    properties:
      createTime:
        allOf:
        - $ref: '#/definitions/util.HTime'
        description: 创建时间
      id:
        description: ID
        type: integer
      postCode:
        description: 岗位编码
        type: string
      postName:
        description: 岗位名称
        type: string
      postStatus:
        description: 状态（1->正常 2->停用）
        type: integer
      remark:
        description: 备注
        type: string
    type: object
  result.Result:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 返回的数据
      message:
        description: 提示信息
        type: string
    type: object
  util.HTime:
    properties:
      time.Time:
        type: string
    type: object
info:
  contact: {}
  description: 后台管理系统API接口文档
  title: 通用后台管理系统
  version: "1.0"
paths:
  /api/captcha:
    get:
      consumes:
      - application/json
      description: 验证码接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 验证码接口
      tags:
      - 系统管理·
  /api/login:
    post:
      description: 用户登录接口
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 用户登录接口
  /api/post/add:
    post:
      description: 新增岗位接口
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.SysPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 新增岗位接口
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"